# perf: only call removeHeader when needed