@gurgunday 2023-08-02
@mcollina can you please make a patch release after this is merged?

https://github.com/fastify/fastify-cookie/commit/54e9bdc7c03b77fe748e82cac7ec0e3ff1d56aea [fixes](https://github.com/fastify/fastify-cookie/commit/54e9bdc7c03b77fe748e82cac7ec0e3ff1d56aea#diff-3ac2d4ddc8bffb36c90b08289c5918425cdb4d38f715ca4adb493d73743cce7dR1191) a case where you set more than 1 cookie after the hook runs